Moved away from explicitly listing nargs, params, and args
This resolves #8 by using lots of macros to calculate all the
required parameters instead of generating them via the python
script. Note that this requires a change in how signatures are
specified in tinycli_regiser(). Instead of a single token with
a single character per parameter, we now use a list of parameters
where each parameter can by multiple characters.

For example:

   tinycli_regiser("start", startFunc, iii)

becomes:

    tinycli_regiser("start", startFunc, i,i,i)